requirements/testing.txt
pytest==8.2.2
pytest-django==4.8.0
pytest-xdist==3.5.0